Home && Cart page Now are Work Properly
Now Can add book to cart and choose quantity of book that user want to buy. And when you click at Cart Icon on navbar. it will navigate you to cart page that show you all info about book you need to buy. The next step is  make  Cart icon on navbar show number of quantity that user add it in the cart. And the next step is make Submit button for user to buy the books that they had add it before.